+++ doemaarwatt/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.11] - 2026-09-02
+
+### Fixed
+
+- Changed default schedule resolution to 15 minutes, as per https://github.com/mauritshartman/RUDE_energy/issues/33
+- Fix for issue https://github.com/mauritshartman/RUDE_energy/issues/23 where PBapp can exceed battery charge/discharge limits
+
 ## [1.1.10] - 2026-07-29
 
 ### Fixed